New static method and bug fix
Added a new static method, sort_dir, that sorts a directory of DICOM files into sub-directories based on the get_dir_from_dicom method output.  More development is needed for this method.

Also, fixed a bug in the show method. Namely, the matplotlib.pyplot package needed to imported directly (instead of trying to reference it through the matplotlib import).