New functionality - mkdir_dicom
Added an option, 'useAllFiles', to the mkdir_dicom stativ method. This allows the user to move a directory of DICOM files one at a time instead of assuming that the parent directory of multiple files is unique. In other words, there are a number of circumstances when multiple directories will exist that contain files that will generate the same directory name.