Bug fixes
Fixed a bug in the mkdir_dicom method. The bug was caused by using 'purebasename' instead of 'basename' to generate the new file name. For files containing an extensions (e.g., file names generated from a UID), the last digits of the file were interpreted as an extensions and removed in the 'purebasename' attribute.

Also fixed an error in the logic of mkdir_dicom. By using the 'ensure_dir' method before verifying the value of 'useAllFiles', a non-existent directory would be generated and a call to the 'rename' method would result in an error. The error would result because the file already exists.